Extract useCreateCryptoIntent
