+++ src/components/donation/Steps/Submit/Crypto/TxSubmit/TxSubmit.tsx
@@ -1,22 +1,17 @@
 import Icon from "components/Icon/Icon";
 import { ErrorStatus, LoadingStatus } from "components/Status";
-import { chains } from "constants/chains";
-import { useErrorContext } from "contexts/ErrorContext";
 import { humanize } from "helpers";
 import { useEffect, useState } from "react";
-import {
-  useConfirmCryptoIntentMutation,
-  useCreateCryptoIntentMutation,
-} from "services/apes";
+import { useConfirmCryptoIntentMutation } from "services/apes";
 import { CryptoSubmitStep } from "slices/donation";
 import { sendDonation } from "slices/donation/sendDonation";
 import { useSetter } from "store/accessors";
-import { CryptoDonation } from "types/aws";
 import { ConnectedWallet } from "types/wallet";
 import ContinueBtn from "../../../common/ContinueBtn";
 import { EstimateStatus, isSuccess } from "../types";
 import { estimateDonation } from "./estimateDonation";
 import { txPackage } from "./txPackage";
+import useCreateCryptoIntent from "./useCreateCryptoIntent";
 
 type Props = {
   classes?: string;
@@ -26,12 +21,8 @@ type Props = {
 export default function TxSubmit({ wallet, donation, classes = "" }: Props) {
   const dispatch = useSetter();
   const [estimate, setEstimate] = useState<EstimateStatus>();
-  const [transactionId, setTransactionId] = useState<string>();
 
-  const { handleError } = useErrorContext();
-
-  const { details, recipient, liquidSplitPct, tip = 0, donor } = donation;
-  const chain = chains[details.chainId.value];
+  const { details } = donation;
   const sender = wallet?.address;
   useEffect(() => {
     if (!sender) return setEstimate(undefined);
@@ -41,48 +32,9 @@ export default function TxSubmit({ wallet, donation, classes = "" }: Props) {
     );
   }, [sender, details]);
 
-  const [createCryptoIntent] = useCreateCryptoIntentMutation();
   const [confirmCryptoIntent] = useConfirmCryptoIntentMutation();
 
-  useEffect(() => {
-    (async () => {
-      try {
-        if (!wallet?.address) {
-          return;
-        }
-
-        const payload: CryptoDonation = {
-          amount: +details.token.amount,
-          tipAmount: tip,
-          chainId: chain.id,
-          chainName: chain.name,
-          denomination: details.token.symbol,
-          splitLiq: liquidSplitPct,
-          walletAddress: wallet.address,
-          endowmentId: recipient.id,
-          source: details.source,
-          donor,
-        };
-        const id = await createCryptoIntent(payload).unwrap();
-        setTransactionId(id);
-      } catch (err) {
-        handleError(err);
-      }
-    })();
-  }, [
-    details.token.amount,
-    details.token.symbol,
-    details.source,
-    recipient.id,
-    liquidSplitPct,
-    chain.id,
-    chain.name,
-    tip,
-    donor,
-    wallet?.address,
-    createCryptoIntent,
-    handleError,
-  ]);
+  const transactionId = useCreateCryptoIntent(donation, wallet);
 
   return (
     <div className={`${classes} grid w-full gap-y-2`}>
